Refactor test suite for data models (#327)
* refactor tests to care for invalids

* split original test_domain.py

* remove orignal test_domain.py

* tidy up test_constraints.py

* tidy-up of test_features

* tidy-up test_priors.py

* tidy-up test_kernels.py

* tidy-up test_molecular and test_molfeatures

* move test_samplers.py

* split filters from base

* remove duplicate test_util.py

* move test_domain_validators.py

* rename and move test_nchoosek_combinatorics.py

* update test pipeline
