+++ .github/workflows/test.yaml
@@ -21,18 +21,7 @@ jobs:
       - name: Minimal Bofire, Python ${{ matrix.python-version }}
         run: pip install "." pytest
       - name: Run domain-only test, Python ${{ matrix.python-version }}
-        run: |
-          pytest tests/bofire/data_models/test_domain.py \
-            tests/bofire/data_models/test_base.py \
-            tests/bofire/data_models/test_constraint_fulfillment.py \
-            tests/bofire/data_models/test_constraints.py \
-            tests/bofire/data_models/test_domain_validators.py \
-            tests/bofire/data_models/test_features.py \
-            tests/bofire/data_models/test_filters.py \
-            tests/bofire/data_models/test_nchoosek_combinatorics.py \
-            tests/bofire/data_models/test_numeric.py \
-            tests/bofire/data_models/test_unions.py \
-            tests/bofire/data_models/test_util.py
+        run: pytest tests/bofire/data_models
 
   testing:
     runs-on: ubuntu-latest
