+++ bofire/data_models/kernels/continuous.py
@@ -1,5 +1,7 @@
 from typing import Literal, Optional
 
+from pydantic import validator
+
 from bofire.data_models.kernels.kernel import Kernel
 from bofire.data_models.priors.api import AnyPrior
 
@@ -19,6 +21,12 @@ class MaternKernel(ContinuousKernel):
     ard: bool = True
     nu: float = 2.5
     lengthscale_prior: Optional[AnyPrior] = None
+
+    @validator("nu")
+    def validate_nu(cls, v, values):
+        if v not in {0.5, 1.5, 2.5}:
+            raise ValueError("nu expected to be 0.5, 1.5, or 2.5")
+        return v
 
 
 class LinearKernel(ContinuousKernel):
