+++ tests/bofire/conftest.py
@@ -3,6 +3,83 @@ from pytest import fixture
 import tests.bofire.data_models.specs.api as specs
 
 
+# invalid fixtures
+@fixture(params=specs.features.invalids)
+def invalid_feature_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.priors.invalids)
+def invalid_prior_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.kernels.invalids)
+def invalid_kernel_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.constraints.invalids)
+def invalid_constraint_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.objectives.invalids)
+def invalid_objective_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.domain.invalids)
+def invalid_domain_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.inputs.invalids)
+def invalid_inputs_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.outputs.invalids)
+def invalid_outputs_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.constraints_container.invalids)
+def invalid_constraints_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.surrogates.invalids)
+def invalid_surrogate_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.acquisition_functions.invalids)
+def invalid_acquisition_function_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.strategies.invalids)
+def invalid_strategy_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.conditions.invalids)
+def invalid_condition_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.outlier_detection.invalids)
+def invalid_outlier_detection_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+@fixture(params=specs.molfeatures.invalids)
+def invalid_molfeatures_spec(request) -> specs.InvalidSpec:
+    return request.param
+
+
+# valid fixtures
 @fixture(params=specs.constraints.valids)
 def constraint_spec(request) -> specs.Spec:
     return request.param
@@ -28,35 +105,51 @@ def acquisition_function_spec(request) -> specs.Spec:
     return request.param
 
 
-try:
-    # in case of the minimal installation these fixtures are not available
-    @fixture(params=specs.strategies.valids)
-    def strategy_spec(request) -> specs.Spec:
-        return request.param
+@fixture(params=specs.inputs.valids)
+def inputs_spec(request) -> specs.Spec:
+    return request.param
 
-    @fixture(params=specs.surrogates.valids)
-    def surrogate_spec(request) -> specs.Spec:
-        return request.param
 
-    @fixture(params=specs.priors.valids)
-    def prior_spec(request) -> specs.Spec:
-        return request.param
+@fixture(params=specs.outputs.valids)
+def outputs_spec(request) -> specs.Spec:
+    return request.param
 
-    @fixture(params=specs.kernels.valids)
-    def kernel_spec(request) -> specs.Spec:
-        return request.param
 
-    @fixture(params=specs.conditions.valids)
-    def condition_spec(request) -> specs.Spec:
-        return request.param
+@fixture(params=specs.constraints_container.valids)
+def constraints_spec(request) -> specs.Spec:
+    return request.param
 
-    @fixture(params=specs.outlier_detection.valids)
-    def outlier_detection_spec(request) -> specs.Spec:
-        return request.param
 
-    @fixture(params=specs.molfeatures.valids)
-    def molfeatures_spec(request) -> specs.Spec:
-        return request.param
+@fixture(params=specs.strategies.valids)
+def strategy_spec(request) -> specs.Spec:
+    return request.param
 
-except AttributeError:
-    pass
+
+@fixture(params=specs.surrogates.valids)
+def surrogate_spec(request) -> specs.Spec:
+    return request.param
+
+
+@fixture(params=specs.priors.valids)
+def prior_spec(request) -> specs.Spec:
+    return request.param
+
+
+@fixture(params=specs.kernels.valids)
+def kernel_spec(request) -> specs.Spec:
+    return request.param
+
+
+@fixture(params=specs.conditions.valids)
+def condition_spec(request) -> specs.Spec:
+    return request.param
+
+
+@fixture(params=specs.outlier_detection.valids)
+def outlier_detection_spec(request) -> specs.Spec:
+    return request.param
+
+
+@fixture(params=specs.molfeatures.valids)
+def molfeatures_spec(request) -> specs.Spec:
+    return request.param
