+++ tests/bofire/data_models/constraints/test_nonlinear.py
@@ -1,9 +1,15 @@
+import importlib
+
 import numpy as np
 import pandas as pd
+import pytest
 
 from bofire.data_models.constraints.api import NonlinearInequalityConstraint
 
+SYMPY_AVAILABLE = importlib.util.find_spec("sympy") is not None
 
+
+@pytest.mark.skipif(not SYMPY_AVAILABLE, reason="requires rdkit")
 def test_nonlinear_constraints_jacobian_expression():
     constraint0 = NonlinearInequalityConstraint(
         expression="x1**2 + x2**2 - x3", features=["x1", "x2", "x3"]
