+++ tests/bofire/data_models/domain/test_domain_nchoosek_combinatorics.py
