+++ tests/bofire/data_models/specs/priors.py
@@ -1,5 +1,7 @@
 import random
 
+from pydantic.error_wrappers import ValidationError
+
 import bofire.data_models.priors.api as priors
 from tests.bofire.data_models.specs.specs import Specs
 
@@ -12,6 +14,17 @@ specs.add_valid(
         "scale": random.random(),
     },
 )
+
+for scale in [-1.0, 0.0]:
+    specs.add_invalid(
+        priors.NormalPrior,
+        lambda: {
+            "loc": random.random(),
+            "scale": scale,  # noqa: B023
+        },
+        error=ValidationError,
+    )
+
 specs.add_valid(
     priors.GammaPrior,
     lambda: {
@@ -19,3 +32,14 @@ specs.add_valid(
         "rate": random.random(),
     },
 )
+
+for rate in [-1.0, 0]:
+    for concentration in [-5, 6]:
+        specs.add_invalid(
+            priors.GammaPrior,
+            lambda: {
+                "concentration": concentration,  # noqa: B023
+                "rate": rate,  # noqa: B023
+            },
+            error=ValidationError,
+        )
