+++ tests/bofire/priors/test_mapper.py
